+++ src/features/auth/reset-password/actions/resetPasswordAction.constants.ts
@@ -1,0 +1,2 @@
+export const RESET_PASSWORD_GLOBAL_ERROR_MESSAGE =
+  "비밀번호를 변경하지 못했습니다. 잠시 후 다시 시도해주세요.";
