+++ src/features/auth/reset-password/actions/resetPasswordActionState.ts
@@ -1,0 +1,19 @@
+export type ResetPasswordActionState =
+  | {
+      status: "idle";
+    }
+  | {
+      status: "field_error";
+      fieldErrors: {
+        password?: string[];
+        confirmPassword?: string[];
+      };
+    }
+  | {
+      status: "global_error";
+      message: string;
+    };
+
+export const initialResetPasswordActionState: ResetPasswordActionState = {
+  status: "idle",
+};
